Save the tasks in hard disk
When the chatbot starts up for the first time, a directory with a text
file, handsome.txt will be created and tasks will be saved automatically
in the file whenever the task list changes. Future starting up of the
bot will load the data from the file as task list.